Treble 10 - RHEI - Add endpoint to handle accessibility form submission
Enable handling incoming JSON data with the body parser
Add factory to build an email transporter with email address and password from the process' environment. Add dotenv package to read environment file and include it in the process.env
Send email when accessibility form endpoint is hit. Respond with error if sending the email fails.
Short circuit and respond with 400 if the accessibility issues form data is missing the description of the issue.
When form is missing user's name or contact information, use default values in the email